backend/requirements.txt
fastapi>=0.111.0
uvicorn>=0.30.1
python-multipart>=0.0.9
langchain>=0.2.14
langchain-community>=0.2.12
langchain-openai>=0.1.22
langchain-huggingface>=0.0.3
langchain-qdrant>=0.1.3
langchain-text-splitters>=0.2.2
qdrant-client>=1.11
sentence-transformers>=3.0
pypdf>=4.3
rank-bm25>=0.2.2
tiktoken>=0.7
docx2txt>=0.8
# v3.1 additions
ragas>=0.1.19
datasets>=2.20.0
pandas>=2.2.0
bcrypt>=4.1.0
PyJWT>=2.8.0
